// src/app/project/models/updateparentsourcerequest.ts
export class UpdateParentSourceRequest {
    sourceId: number;
    updateFrequency: string;
    sourceName: string;
    sourceNo: number;
    dbSourceNO: number;
    ownerId: number;
    publicStatusCommnent: string;
    frequencyId: number;
    sourceTypeId: number;
    updateType: string;
    processedAt: string;
    isActive: any;
    hasPhoto: any;
    photoPath: string;
    description: string;
    notes: string;
}